requirements.txt
seaborn~=0.12.2
numpy~=1.24.3
pandas~=2.0.0
scikit-learn~=1.2.2
plotly~=5.14.1
dash~=2.9.3
dash-daq
sqlalchemy~=2.0.10
python-dotenv
psycopg2
dill
-e .